xen: Add helper function to switch VCPU
The information about whose Xen VM vcpu's stack to unwind is carried in
struct UXEN_info, because we need to know it while running the
xen-specific stuff, but have no way to pass it directly from the calling
code due to the fixed signatures of the libunwind wrapper functions.
There is nothing special about this information that prevents us from
changing it when we want (though doing it in the middle of a stack
unwind is probably a quite bad idea). So _UXEN_change_cpu gives us a
helper function to change the vcpu from outside, without exposing all of
struct UXEN_info to every application that links to us.